% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCoverage.R
\name{coverageSummary}
\alias{coverageSummary}
\title{Reads a mod.df object created by \code{\link{readModDir}} with ucov/tcov.}
\usage{
coverageSummary(mod.row, method)
}
\arguments{
\item{method}{How is 'coverage' summarized. "mean", "median", "max", "q95". Default: `mean'}

\item{mod.df}{A modification data.frame object}

\item{set.na}{In log2FC field, coverts all NA and NaN values to <VALUE> [FALSE]. Set to FALSE to leave unchanged.}

\item{set.inf}{In log2FC field, coverts all Inf or -Inf values to <VALUE> [FALSE]. Set to FALSE to leave unchanged}
}
\value{
A data.frame of <Gene>, <uCov_summary>, <tCov_summary>
}
\description{
Reads a mod.df object created by \code{\link{readModDir}} with ucov/tcov.
}
\examples{
geneCov.wt <- geneCoverage( mod.df.wt1, method = "mean")
geneCov.ko <- geneCoverage( mod.df.ko1, method = "mean")
}
\keyword{RNAframework}
\keyword{rf-modcall}
