% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCovRun.R
\name{covrunSummary}
\alias{covrunSummary}
\title{Reads a mod.df object created by \code{\link{readModDir}} with ucov/tcov for coverage run}
\usage{
covrunSummary(mod.row, threshold = 0, col.select = 8)
}
\arguments{
\item{threshold}{The number of reads in coverage required for a run count to start}

\item{mod.df}{A modification data.frame object}
}
\value{
The maximum length (nt) of consecutive bases with coverage greater then threshold
}
\description{
Reads a mod.df object created by \code{\link{readModDir}} with ucov/tcov for coverage run
}
\examples{
geneCov.wt <- geneCovRun( mod.df.wt1, theshold = 5 )
}
\keyword{RNAframework}
\keyword{rf-modcall}
