% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiveSigma.R
\name{fiveSigma}
\alias{fiveSigma}
\title{Mean +/- 5*sd of a numerical vector}
\usage{
fiveSigma(numVec, greater = T, dropZero = T, verbose = F)
}
\arguments{
\item{numVec}{numerical vector}

\item{greater}{return(mean + 5sd). Else return(mean - 5sd). Default = T}

\item{dropZero}{exclude values of zero from the calulation of mean/sd}

\item{verbose}{prints values}
}
\value{
Numeric
}
\description{
Used for score and ratio cut-off thesholding.
}
\examples{
fiveSigma( c(10,0,0,15,10) )

# Verbose output
fiveSigma(c(10,0,0,0,10,15), verbose = T)
#"The mean of the vector is: 11.6666666666667"
#"The standard deviation of the vector is: 2.88675134594813"
#"Five standard deviations from mean is: 26.1004233964073"
#26.10042
}
\keyword{RNAframework}
\keyword{rf-modcall}
