% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getORF.R
\name{getORF}
\alias{getORF}
\title{Wrapper function for EMBOSS 'getorf' program}
\usage{
getORF(ntFa, orfFa)
}
\arguments{
\item{ntFa}{A transcriptome reference file}

\item{orfFa}{An ORF output file generated by getorf.
Default: "getorf reference.fa refORF.fa -find 1 -noreverse"}
}
\value{
orf.df object. A data.frame describing the ORFs found
}
\description{
Wrapper function for EMBOSS 'getorf' program
}
\examples{
# Run getorf
system('getorf fa/reference.fa fa/refORF.fa -find 1 -noreverse')
orf.df = getORF(ntFa = 'reference.fa', orfFa = 'reference_orf.fa')
}
\keyword{RNAframework}
\keyword{getorf}
\keyword{rf-modcall}
