% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectNT.R
\name{intersectNT}
\alias{intersectNT}
\title{Find overlap between two modNT.df objects}
\usage{
intersectNT(nt.df1, nt.df2)
}
\arguments{
\item{mod.df1}{First modification data.frame object}

\item{mod.df2}{Second modification data.frame object}
}
\value{
A list describing the intersection between mod.df1 and mod.df2
}
\description{
Find overlap between two modNT.df objects
}
\examples{
RTstop.wt1 <- sigMod( mod.df.wt1, method = "fiveSigma", flank.seq = 5)
RTstop.bg.wg1 <- sigMod( mod.df.wt1, method = "fiveSigma", flank.seq = 5, randomize = T)
}
\keyword{RNAframework}
\keyword{rf-modcall}
