% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modToFa.R
\name{modToFa}
\alias{modToFa}
\title{Write a Fasta file for each flank.seq in a nucleotide.modification data.frame}
\usage{
modToFa(file, modNT.df, overwrite = T)
}
\arguments{
\item{file}{fasta file to write to}

\item{modNT.df}{nucleotide.modification data.frame (either significant or control)}

\item{overwrite}{if file exists, delete it and overwrite. Else append to file. Default = T}
}
\value{
NULL writes to file
}
\description{
Write a Fasta file for each flank.seq in a nucleotide.modification data.frame
}
\examples{
modToFa('RTstop.motif.fa', RTstop.wt1)
# Create a background set x100 for motif analysis by HOMER

for (N in 1:100){
rand.modNT.N <- sigMod( RTstop.wt1, method = "fiveSigma", flank.seq = 5, randomize = T)
  modToFa( 'BG.motif.fa', rand.modNT.N, overwrite = F)
}
}
\keyword{RNAframework}
\keyword{rf-modcall}
