% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReps.R
\name{plotReps}
\alias{plotReps}
\title{Plot Score vs. Score of two mod.df object
Will also draw vertical/horizontal line at \code{\link{fiveSigma}}}
\usage{
plotReps(mod.df1, mod.df2, log = T)
}
\arguments{
\item{mod.df1}{X-axis modification data.frame}

\item{mod.df2}{Y-axis modification data.frame}

\item{log}{Plot log(score, base = 10) instead of linear. Default = T}
}
\value{
geom_hex() object
}
\description{
Plot Score vs. Score of two mod.df object
Will also draw vertical/horizontal line at \code{\link{fiveSigma}}
}
\examples{
plot <- plotReps(RTstop.wt1, RTstop.wt2)
plot <- plot + xlab('log(RT stop score)') + ylab('ratio')
plot <- plot + ggtitle('RT stop score between replicates')
plot

}
\seealso{
\code{\link{geom_hex}}
}
\keyword{RNAframework}
\keyword{rf-modcall}
