% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScore.R
\name{plotScore}
\alias{plotScore}
\title{Plot Score vs. Ratio for each nt in mod.df}
\usage{
plotScore(mod.df, log = T)
}
\arguments{
\item{mod.df}{modification data.frame object}

\item{log}{Plot log(score, base = 10) instead of linear. Default = T}
}
\value{
geom_hex() object
}
\description{
Plot Score vs. Ratio for each nt in mod.df
}
\examples{
plot <- plotScore(RTstop.wt1)
plot <- plot + xlab('log(RT stop score)') + ylab('ratio')
plot

}
\seealso{
\code{\link{geom_hex}}
}
\keyword{RNAframework}
\keyword{rf-modcall}
