% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScoreRatio_bar.R
\name{plotScoreRatio_Bar}
\alias{plotScoreRatio_Bar}
\title{Plot box-plot for Score/Ratio of each nt in mod.df}
\usage{
plotScoreRatio_Bar(mod.df, add.jitter = T, log.scale = T)
}
\arguments{
\item{mod.df}{modification data.frame object}

\item{add.jitter}{Plot log(score, base = 10) instead of linear. Default = T}

\item{log.scale}{Plot log(score, base = 10) instead of linear. Default = T}
}
\value{
ggplot2 object
}
\description{
Plot box-plot for Score/Ratio of each nt in mod.df
}
\examples{
plot <- plotScoreRatio_bar(RTstop.wt1, add.jitter = F)
plot <- plot + xlab('log(Score / RT stop score)')
plot

}
\seealso{
\code{\link{geom_boxplot}}
}
\keyword{RNAframework}
\keyword{rf-modcall}
