% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readModXML.R
\name{readModXML}
\alias{readModXML}
\title{Reads a rf-modcall XML files}
\usage{
readModXML(file, na.asZero = T)
}
\arguments{
\item{file}{XML file to read}

\item{na.asZero}{Convert NA in rf-modcall (score and ratio) to zreo. Default: T}
}
\value{
List object for each transcript containing scores/sequence for Transcript
}
\description{
Reads a rf-modcall XML files
}
\examples{
readModXML(file = 'rf-mod.experiment/PTEN.xml')
}
\seealso{
\code{\link{readModDir}} Wraps thus function across an entire directory
}
\keyword{RNAframework}
\keyword{rf-modcall}
