% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPeakBed.R
\name{readPeakBed}
\alias{readPeakBed}
\title{Reads an output bed files from "rf-peakcall"}
\usage{
readPeakBed(bed.path, as.GRanges = F)
}
\arguments{
\item{as.GRanges}{Returns the bed file as a "GRanges object"}

\item{inbed}{Bed file input generated from "rf-peakcall"}
}
\value{
Returns a bed file as as data.frame or GRanges object
}
\description{
Reads an output bed files from "rf-peakcall"
}
\examples{
wt1.bed <- readPeakBed(bed.path = "primaseq/data/wt1.bed")
}
\keyword{RNAframework}
\keyword{rf-peakcall}
