% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigMod.R
\name{sigMod}
\alias{sigMod}
\title{Reads a mod.df object created by \code{\link{readModDir}}}
\usage{
sigMod(mod.df, method = "fiveSigma", score.cutoff, ratio.cutoff,
  randomize = F, flank.seq = 5)
}
\arguments{
\item{mod.df}{A modification data.frame object}

\item{method}{Default = "fiveSigma".
\code{\link{fiveSigma}}: Thresholds are set as the mean non-zero score|ratio
plus five standard-deviations of all nucleotides in the transcriptome.
This is making the assumption that the vast majority of NT are non-reactive
and returns only exceptional scores.}

\item{score.cutoff}{Return nucleotides with a score greater than cutoff}

\item{ratio.cutoff}{Return nucleotides with a ratio greater than cutoff}

\item{randomize}{For each significant NT, return a random NT in that transcript instead.
This is useful for creating a background set of sequences based on transcriptome}

\item{flank.seq}{For each NT, return +/- this many flanking NT in the flank.seq column}
}
\value{
Returns a significant.nucleotide.modification data.frame
}
\description{
Reads a mod.df object created by \code{\link{readModDir}}
}
\examples{
RTstop.wt1 <- sigMod( mod.df.wt1, method = "fiveSigma", flank.seq = 5)
RTstop.bg.wg1 <- sigMod( mod.df.wt1, method = "fiveSigma", flank.seq = 5, randomize = T)
}
\keyword{RNAframework}
\keyword{rf-modcall}
